# openapi/v0.yml
openapi: '3.1.0'
info:
  title: TwextHub API
  version: '0.1.0'
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    TwextHub is a lightweight registry for publishing and discovering
    Twext-compiled TurboWarp extensions ("Twexts"). Packages are namespaced
    to their publishing account (`@namespace/id`) and versioned with SemVer.
    Because the Twext compiler does not embed manifest metadata into its
    compiled output (unlike the old Warp/Fluorite `Warp.meta` convention),
    every publish call must submit its manifest fields explicitly as JSON
    alongside the compiled code — TwextHub never parses the JS blob to
    recover metadata.

    Every authenticated (`Userspace`/`Adminspace`) endpoint requires the
    caller to have accepted the current Terms of Service; a caller who
    hasn't gets a `403` (see `Forbidden`) until they call
    `POST /terms/accept`.

    Bearer tokens come in two kinds. **Sessions** are created by
    `/auth/login` or `/auth/signup`, expire after 7 days, and implicitly
    satisfy every scope. **Automation tokens** are created explicitly via
    `POST /tokens`, are scoped to just `publish` and/or `yank` for CI use
    (e.g. a `twext publish` GitHub Action), and never expire unless the
    caller sets one. Only a session may manage sessions or automation
    tokens, act on `/admin/*` endpoints, or change account settings —
    an automation token calling any of those gets a `403`.
servers:
...
    variables:
...
security:
  - BearerAuth: []
paths:
...
  /tokens:
    get:
      summary: List a caller's automation tokens
      operationId: listAutomationTokens
      tags: [Userspace]
      description: Session-authenticated only. `namespace` defaults to the caller; an admin may pass it to inspect another account's tokens. Never returns token secrets.
      parameters:
        - name: namespace
          in: query
          schema: { $ref: '#/components/schemas/Namespace' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated automation tokens
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AutomationToken' }
                  pagination: { $ref: '#/components/schemas/CursorPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      summary: Create an automation token
      operationId: createAutomationToken
      tags: [Userspace]
      description: Session-authenticated only. The `token` secret is returned exactly once, in this response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name: { type: string, maxLength: 80, description: 'e.g. "CI publish".' }
                scopes:
                  type: array
                  items: { $ref: '#/components/schemas/Scope' }
                  minItems: 1
                expiresInDays:
                  { type: integer, minimum: 1, description: Omit for a token that never expires. }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationTokenCreated' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /tokens/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    patch:
      summary: Rename or rescope an automation token
      operationId: updateAutomationToken
      tags: [Userspace]
      description: Self or admin, session-authenticated only. Never returns or reissues the token secret.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name: { type: string, maxLength: 80 }
                scopes:
                  type: array
                  items: { $ref: '#/components/schemas/Scope' }
                  minItems: 1
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AutomationToken' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      summary: Revoke an automation token
      operationId: deleteAutomationToken
      tags: [Userspace]
      description: Self or admin, session-authenticated only.
      responses:
        '204': { description: Revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------------- Users
  /users:
    get:
      summary: List accounts
      operationId: listUsers
      tags: [Open]
      security: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated accounts
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  pagination: { $ref: '#/components/schemas/CursorPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /users/{namespace}:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
    get:
      summary: Get an account
      operationId: getUser
      tags: [Open]
      security: []
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update an account
      operationId: updateUser
      tags: [Userspace]
      description: >
        Self or admin. Only an admin may set `role`. Changing `password`
        revokes all of that account's existing tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                displayName: { type: string, maxLength: 80 }
                password: { type: string, minLength: 8, writeOnly: true }
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
    delete:
      summary: Delete an account
      operationId: deleteUser
      tags: [Userspace]
      description: Self or admin. Cascades to all of that account's extensions and versions.
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------- Discovery
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque pagination cursor from a previous response.
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 50 }
    NamespacePath:
      name: namespace
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Namespace' }
...
  schemas:
    Namespace:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
...
    Role:
      type: string
      enum: [admin, normal]
    Scope:
      type: string
      enum: [publish, yank]
      description: Only meaningful for automation tokens — sessions implicitly satisfy every scope.
...
    AutomationToken:
      type: object
      required: [id, name, scopes, createdAt, expiresAt, lastUsedAt]
      properties:
        id: { type: string }
        name: { type: string }
        scopes:
          type: array
          items: { $ref: '#/components/schemas/Scope' }
        createdAt: { type: string, format: date-time }
        expiresAt:
          {
            type: ['string', 'null'],
            format: date-time,
            description: '`null` if the token never expires.',
          }
        lastUsedAt: { type: ['string', 'null'], format: date-time }
    AutomationTokenCreated:
      allOf:
        - $ref: '#/components/schemas/AutomationToken'
        - type: object
          required: [token]
          properties:
            token:
              {
                type: string,
                description: The bearer token secret. Shown only in this one response.,
              }
    User:
      type: object
      required: [namespace, displayName, role, hasPublished, createdAt, termsAcceptedVersion]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        hasPublished: { type: boolean }
        createdAt: { type: string, format: date-time, readOnly: true }
        termsAcceptedVersion:
          type: ['integer', 'null']
          description: Terms of Service `version` this account last accepted, or `null` if never accepted.
...
    CursorPage:
      type: object
      required: [nextCursor, hasMore]
      properties:
        nextCursor: { type: ['string', 'null'] }
        hasMore: { type: boolean }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          { type: string, format: uri, example: 'https://hub.twext.dev/errors/validation-error' }
        title: { type: string, example: 'Validation Error' }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
    BadRequest:
      description: Invalid request
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unauthorized:
      description: Missing or invalid bearer token
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: >
        Authenticated, but not allowed to perform this action — e.g. not the
        resource owner or an admin, a namespace mismatch on publish, the
        current Terms of Service haven't been accepted yet
        (see `POST /terms/accept`), an automation token missing the
        required scope, or an automation token used against a
        session-only endpoint (session/token management, `/admin/*`,
        account settings).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Resource not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    UnprocessableEntity:
      description: Semantically invalid input (e.g. non-SemVer version, id mismatch)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer